requirements.txt
Babel==2.6.0
beautifulsoup4==4.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cogwheels==0.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-modelcluster==4.3
django-oscar==1.6.6
django-oscar-wagtail==0.1.3
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-taggit==0.23.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
draftjs-exporter==2.1.5
factory-boy==2.11.1
Faker==1.0.2
html5lib==1.0.1
idna==2.8
mock==2.0.0
pbr==5.1.1
phonenumbers==8.10.4
Pillow==5.4.1
purl==1.4
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
wagtail==2.3
wagtailmenus==2.12
webencodings==0.5.1
Willow==1.1
